add streamlit app to CLI, reformat app code
